move quantum program out of the layout module, add attribute qubit mapping
after discussing the design with the maintainer:
- layout/ is only for laying out types, so QuantumInstr and QuantumProgram
  move into a new program/ module - the quantum layout and its cache stay
  under layout/quantum/
- add QuantumProgram::attr_qubits(attr) -> (offset, width) so the Q3L
  emitter knows which qubits each attribute maps to, for measurement
- ancillas are summed for the total (not freed), per the discussion
- arrays and wiring into the cast pass left as follow-ups
